requirements.txt
langchain>=0.1.16
langchain-community>=0.0.32
langchain-huggingface>=0.0.3
transformers>=4.38.0
sentence-transformers>=2.6.1
huggingface-hub>=0.22.2
torch>=2.1.0
accelerate>=0.27.2
faiss-cpu>=1.7.4
streamlit>=1.32.0
python-dotenv>=1.0.1
tiktoken>=0.6.0
numpy>=1.24.0
pandas>=2.0.0